Merge qt to develop

See merge request students/CPP4_3DViewer_v2.0.ID_353537/Team__TL__donnapoc_student.21_school.ru_.87c36bc4_968b_4a46/CPP4_3DViewer_v2.0-1!1

+++ src/parser/parser.hpp
@@ -3,8 +3,10 @@
 #include <fstream>
 #include <iostream>
 #include <string>
+#include <locale>
 
 #include "../figure/figure.hpp"
+
 using namespace s21;
 
 namespace s21 {
@@ -17,8 +19,8 @@ class Parser {
  private:
   Figure& figure;
   long v = 0;
-  vector<long> polygon_line;
-  vector<long> sub_line;
+  vector<unsigned int> polygon_line;
+  vector<unsigned int> sub_line;
 
   int num_slash = 0, prev_num_slash = 0;
   int pattern_size = 0, prev_pattern_size = 0;
